requirements.txt
fastapi
uvicorn
hypercorn
daphne
requests
psycopg2-binary
fpdf   
email-validator          
python-multipart
apscheduler
aiosmtplib
python-dotenv